Add tests for batched note decryption.
